Refactor: Adicioando filtro de exceção para quando a Pizza não existe

+++ src/pizzas/pizzas.service.ts
@@ -45,6 +45,9 @@ export class PizzasService {
   }
 
   async delete(id: number) {
+    const orderItem = await this.pizzaRepository.findOneBy({ id });
+    if (!orderItem) throw new NotFoundException('Pizza não encontrada');
+
     await this.pizzaRepository.delete(id);
     return { msg: 'Pizza deletada com sucesso!' };
   }
